Update comments, organization
